+++ ci/cleanup.sh
@@ -1,1 +1,0 @@
-./kafka_test_teardown.sh